(pasteimage) FS#4750: Fix imgs not pasted at caret pos in Chrome
Because images can be fetched from the Chrome clipboard another
paste handler is used for Chrome.

On CTRL+V the caret is redirected to an invisible field where
the paste is made. This is so that the paste content can be
processed before put in the real text.

The function editor.insertElement() was used to insert the new
img element however because of the different paste handler for
Chrome at the time this function was called the caret was still
inside the hidden pastebin which meant the new image was always
inserted at the beginning of the document.

To solve this the plugin now listens for beforePaste and at that
time saves the current caret position and inserts the img element
at that saved position.