CCC-49 RTF support for chat
added code to check for CKEDITOR config imageMaxSize. This config
value (in bytes) will be used to check if image file included
from the dialog window has exceeded the maximum configured value.
In cention chat, this configuration value is set by the
chat-max-image-size value. This code change is not part of
this plugin's main branch in GIT. From here on, whenever we need
to refresh the plugin code with fixes from the main git branch,
we'll need to port this fix back