Avoid cursor moved when cention-emoji replace emoji
When emoji is replaced, the cursor position most likely get out
of range because the emoji number of characters became less. And
most probably this is the reason the cursor position is default
to either right or left side of the replaced emoji char. This
commit calcalute the new cursor position everytime an emoji is
replaced.