Fix last emoji change event not fire and add emoji replace inline
There was a bug where last typed characters are replaced with
emoji then this replacement will not reflect as event change. It
was caused the reason the replacement way is not go through the
Ckeditor text insertion but it go through direct DOM manipulation
as the emoji replacement was modified from emojify library.

Ckeditor inline has slight difference on the DOM return from
editor instance object compare to classic Ckeditor.
editor.document suitable for classic as it directly pointing to
the text body of editable input. But editor.document is not
suitable for Ckeditor inline because it pointing to the current
windows.document object. Instead ckeditor.container will be better
choice to be used as input for emoji search and replace function.
The plugin code now need to detect if its editor is inline or not
then choose the better object as parameter.

Emoji DOM replacement now done on plugin.js (Ckeditor plugin side)
as to make the code clearer and isolate plugin from emojify
library. Also added limitation on only focus-ed plugin can has the
auto emoji conversion as there is some bugs on Ckeditor.inline
range select when out of focus. Also isolate the auto convert
option by Ckeditor instance instead of globally as single plugin
option. Switching on/off a Ckeditor instance's cention_emoji auto
conversion option will not affect the other Ckeditor instance on
the same page.

Also adjusted the skin to specifically target the css for Ckeditor
inline so that is has border on left right sides and look square.