bugfix: Unable to get property 'checkReadOnly' of undefined or null reference
when ckeditor first renders its trying insert html in editors editable
area which editor.getSelection(): CKEDITOR.dom.selection -> A selection
object or null if not available for the moment. But if editor's editable
element lost focus(somehow*), this return null. getSelection() return multiple
or null ranges. so if the ranges return null then it occurs error.

getRanges(): Range instances that represent the current selection.

checkReadOnly(): Check if elements at which the range boundaries anchor
are read-only, with respect to contenteditable attribute.