Provide Command To Update Floating Toolbar Position
If inline ckeditor with floating attribute enabled and place
inside DOM that is draggable (such as Bootstrap modal), then the
ckeditor toolbar can not be appeared when focus-ed. Most likely
because the drag event was controlled by Bootstrap. The
workaround for this is to disable dragable on element that
contains the ckeditor. Without drag event, this plugin can not
know when to update the toolbar position again if the modal is
dragged or moved. This commit provide an optional command that
allow user to ask for new calibration of toolbar position. The
command most likely call from drag event of the modal.