Fix CS-350 cention_spellchecker toggle doesnt update ckeditor states
The case hapened after copying the text to ckeditor and customer was
clicking on `Perform spellchecker` button and that was showing the
wrong text red colored and then if we toggle on the spellchecker and not
changing anything from ckeditor body then the ckeditor still keeping the
red color because when the perform spellchecker button clicked then there
are some html tags with css class has injected to the editor to make the
text colored.

It happened because after toggling on `perform spellchecker button` doesn't
update the ckeeditor states in our ckeditor components which causes the problem.

Previously we got the ckeditor's document body while we do save or sent mail
but that has created few more issues such not updating the area's attachment etc.

So now we maintain the ckeditor body from reactjs component level, which is
related to trigger any action on editor. Above cases the editors inside states
updated but our component doesnt trigger any event to let it to update the
component states which causing the issue. When spellchecker toggle happen then
it will fire `focus` event after releasing the spellchecker event.