Fix Shared Space Toolbar Got Hidden When Different Toolbars Are Used
When a toolbar created with sharedspace plugin and later get
destroyed and recreated, it can not be revealed when it got
recreated because the plugin was memorized the toolbar have already
had one visible (which already been destroyed) and thus the later
created toolbar(s) need to be hidden. This caused problem when the
currently visible toolbar is destroyed and recreated. This commits
will not just remembered if the toolbar had already have a visible
toolbar, but also remember which editor is currently visible. If
the currently visible toolbar is same as the one that is requesting
to create a new toolbar, then it should be shown again and should
not be hidden. This can only happen when the toolbar is destroyed
and recreated.