Fix CCC-3496 copy/paste text not working in CKEditor
The problem causes by CCC-3388 however the main problem of duplicate
issue had resolved.
The current problem was happened due to preventDefault happens for all
cases either text or image during copy/paste. Now the preventDefault
happening only to handle the image and text will be as it is under
ckeditor default.